Update documentation relating to network changes
The `autoRefreshOnNetworkChange` flag description has been updated to
remove the warning about it being experimental. This flag is stable and
will be recommended until we remove the reload-on-network-change
behavior completely. The description was also updated to better explain
what the flag does.

The notice on the `networkChanged` event about needing to disable the
default auto-reload behaviour has been updated to reflect our future
plans for this event.