Remove paragraph about default token list (#151)
At the moment the `eth-contract-metadata` repo is swamped with
unreviewed issues and PRs, so effectively we are not accepting new
tokens at the moment. Even though this paragraph does a lot to dissuade
people from submitting their token for inclusion, it still gives a
false impression.

I've removed it for now, but we can restore it again as soon as we have
a plan for how to handle these requests going forward. Hopefully this
is soon.